+++ CHANGES.md
@@ -4,7 +4,7 @@ Changes
 1.10.0 to 1.11.0
 ----------------
 * Personalized radio.
-* (Not finished) Persistant Home button.
+* Persistant Home button.
 * Refactored `buildSearchIndex.php` into into `StreamsSearchIndexer` and updated `buildSearchIndex.php` accordingly. Also added unit tests.
 
 1.9.0 to 1.10.0